refactor: extract PDF structural headings
